backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
psycopg2-binary
python-dotenv
firebase-admin